remove excess .h + navi-> wh
