+++ src/ariadne_wallhug.h
@@ -31,8 +31,23 @@ extern "C" {
 
 struct Thing;
 struct Coord3d;
-struct Navigation;
 struct SlabMap;
+
+struct Navigation {
+  unsigned char navstate;
+  unsigned char side;
+  unsigned char wallhug_retry_counter;
+  unsigned char wallhug_state;
+  unsigned char push_counter;
+  long dist_to_final_pos;
+  long distance_to_next_pos;
+  int32_t angle;
+  SubtlCodedCoords first_colliding_block;
+  SubtlCodedCoords second_colliding_block;
+  PlayerBitFlags owner_flags[2];
+  struct Coord3d pos_next;
+  struct Coord3d pos_final;
+};
 
 enum WallHugSideState {
     WaHSS_Initial = 0,
@@ -51,6 +66,7 @@ SubtlCodedCoords dig_to_position(PlayerNumber plyr_idx, MapSubtlCoord basestl_x,
 TbBool slab_good_for_computer_dig_path(const struct SlabMap *slb);
 short get_hug_side_options(MapSubtlCoord stl1_x, MapSubtlCoord stl1_y, MapSubtlCoord stl2_x, MapSubtlCoord stl2_y, SmallAroundIndex direction, PlayerNumber plyr_idx,
     MapSubtlCoord *ostla_x, MapSubtlCoord *ostla_y, MapSubtlCoord *ostlb_x, MapSubtlCoord *ostlb_y);
+void initialise_wallhugging_path_from_to(struct Navigation *navi, struct Coord3d *mvstart, struct Coord3d *mvend);
 /******************************************************************************/
 #define CHECK_SLAB_OWNER (flag_is_set(crt_owner_flags, to_flag(slabmap_owner(slb)))) // return TRUE if the slab's owner is stored in crt_owner_flags
 #define IGNORE_SLAB_OWNER_CHECK 0 // crt_owner_flags can be set to 0 to nullify the check for the slab's owner
