+++ src/creature_states_train.c
@@ -40,7 +40,6 @@
 #include "room_data.h"
 #include "room_jobs.h"
 #include "map_utils.h"
-#include "ariadne_wallhug.h"
 #include "gui_soundmsgs.h"
 #include "game_legacy.h"
 #include "post_inc.h"
