+++ src/creature_states_wrshp.c
@@ -38,7 +38,6 @@
 #include "room_data.h"
 #include "room_jobs.h"
 #include "map_utils.h"
-#include "ariadne_wallhug.h"
 #include "gui_soundmsgs.h"
 
 #include "game_legacy.h"
