+++ src/player_utils.c
@@ -46,7 +46,6 @@
 #include "config_terrain.h"
 #include "map_blocks.h"
 #include "map_utils.h"
-#include "ariadne_wallhug.h"
 #include "game_saves.h"
 #include "game_legacy.h"
 #include "frontend.h"
