+++ src/spdigger_stack.c
@@ -49,7 +49,6 @@
 #include "power_hand.h"
 #include "map_utils.h"
 #include "map_events.h"
-#include "ariadne_wallhug.h"
 #include "gui_soundmsgs.h"
 #include "front_simple.h"
 #include "game_legacy.h"
