+++ src/thing_doors.c
@@ -31,8 +31,6 @@
 #include "thing_effects.h"
 #include "config_terrain.h"
 #include "creature_senses.h"
-#include "ariadne.h"
-#include "ariadne_wallhug.h"
 #include "map_blocks.h"
 #include "map_ceiling.h"
 #include "map_utils.h"
